+++ tests/testthat/test-etl.R
@@ -9,16 +9,27 @@ test_that("sqlite works", {
   expect_true(file.exists(find_schema(cars_sqlite)))
   expect_message(find_schema(cars_sqlite, "my_crazy_schema", "etl"))
   expect_output(summary(cars_sqlite), "files")
-  expect_message(cars_sqlite %>% etl_create(), "Uploading")
-  expect_message(cars_sqlite %>% etl_init(), "Loading SQL script")
+  expect_message(cars_sqlite %>% etl_create(), "Loading")
+  expect_message(cars_sqlite %>% etl_init(), "Running SQL script")
   expect_message(
     cars_sqlite %>% etl_cleanup(delete_raw = TRUE, delete_load = TRUE),
     "Deleting files")
 })
 
+test_that("default works", {
+  dplyr_sqlite <- etl("dplyr")
+  expect_s3_class(dplyr_sqlite, c("etl_dplyr", "etl", "src_sqlite", "src_dbi"))
+  expect_output(summary(dplyr_sqlite), "files")
+  expect_message(dplyr_sqlite %>% etl_update(), "Loading")
+  expect_message(
+    dplyr_sqlite %>% etl_cleanup(delete_raw = TRUE, delete_load = TRUE),
+    "Deleting files")
+})
+
+
 test_that("dplyr works", {
   expect_message(cars <- etl("mtcars") %>%
-    etl_create(), regexp = "Uploading")
+    etl_create(), regexp = "Loading")
   expect_gt(length(src_tbls(cars)), 0)
   tbl_cars <- cars %>%
      tbl("mtcars")
@@ -30,7 +41,7 @@ test_that("dplyr works", {
   # double up the data
   expect_message(
     cars %>%
-      etl_update(), regexp = "Uploading")
+      etl_update(), regexp = "Loading")
   res2 <- tbl_cars %>%
     collect()
   expect_equal(nrow(res2), 2 * nrow(mtcars))
@@ -53,7 +64,7 @@ test_that("mysql works", {
 test_that("valid_year_month works", {
   expect_equal(
     nrow(valid_year_month(years = 1999:2001, months = c(1:3, 7))), 12)
-  test_dir <- "~/dumps/airlines"
+#  test_dir <- "~/dumps/airlines"
   # if (require(airlines) & require(etl) & dir.exists(test_dir)) {
   #   airlines <- etl("airlines", dir = test_dir) %>%
   #     etl_extract(year = 1987)
@@ -86,16 +97,25 @@ test_that("smart_download works", {
   cars <- etl("mtcars")
   # first download some files
 #  if (!.Platform$OS.type == "windows") {
+    expect_message(etl_cleanup(cars, pattern = ".", delete_raw = TRUE, delete_load = TRUE), "Deleting")
     urls <- c("https://raw.githubusercontent.com/beanumber/etl/master/etl.Rproj",
               "https://www.reddit.com/robots.txt")
     expect_length(smart_download(cars, src = urls), 2)
     # then try to download them again
-    expect_length(list.files(attr(cars, "raw_dir")), 3)
     expect_length(smart_download(cars, src = urls), 0)
-    expect_message(etl_cleanup(cars, pattern = "com", delete_raw = TRUE), "Deleting")
+    expect_message(etl_cleanup(cars, pattern = ".", delete_raw = TRUE, delete_load = TRUE), "Deleting")
 #  }
 })
 
+
+test_that("cities works", {
+  cities_sqlite <- etl("cities")
+  # fails on check() but not on test()?? issue #37
+#   expect_message(cities_sqlite %>% etl_create(), "Loading")
+  expect_message(
+    cities_sqlite %>% etl_cleanup(delete_raw = TRUE, delete_load = TRUE),
+    "Deleting files")
+})
 
 test_that("MonetDBLite works", {
   if (require(MonetDBLite)) {